#528 FakeOctokit extracted with 100% coverage
